Split the webviewMenu into files from its composing parts
